subrem/requirements.txt
# For mmcv compiled with torch 2.3 + CUDA 12.1
--find-links https://download.openmmlab.com/mmcv/dist/cu121/torch2.3/index.html

accelerate==0.33.0
addict==2.4.0
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
contourpy==1.2.1
controlnet-aux==0.0.9
cupy-cuda12x==13.2.0
diffusers==0.29.2
easyocr==1.7.1
einops==0.8.0
fastrlock==0.8.2
ffmpeg-python==0.2.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
gdown==5.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.1
imageio==2.34.2
importlib-metadata==8.1.0
jinja2==3.1.4
jpeg4py==0.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mmcv==2.2.0
mmengine==0.10.4
mpmath==1.3.0
ninja==1.11.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pandas==2.2.2
pyclipper==1.3.0.post5
pycocotools==2.0.8
pydantic==2.8.2
pydantic-core==2.20.1
pyparsing==3.1.2
python-bidi==0.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.1
regex==2024.5.15
replicate==0.30.1
rich==13.7.1
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.14.0
shapely==2.0.5
sniffio==1.3.1
soupsieve==2.5
sympy==1.13.1
termcolor==2.4.0
tifffile==2024.7.21
timm==0.6.7
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.43.1
tzdata==2024.1
yapf==0.40.2
boto3==1.34.83
